pyproject.toml
[tool.poetry]
name = "watermarker"
version = "0.1.0"
description = ""
authors = ["thorgeir <thorgeirsigurd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "^10.0.0"
click = "^8.1.6"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.0.0"
mypy = "^1.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
